* master: (669 commits)
  Revert "Fix uniboard layout fallback for old browsers"
  add a /dev/cli command to clear disposable email cache
  fix bad auto-import
  nicer code with scala for
  bump study.updatedAt when moving without sync - closes #12882
  fix typos
  code golf
  move some static content out of app/
  fix test
  refactor PicfitApi.bodyImage after #12879
  set ublog image rel
  naming tweaks
  optional MarkdownRender asset domain
  type AssetDomain in MarkdownRender
  New Crowdin updates (#12870)
  lichess forks can play too
  ublog image uploads
  Fix uniboard on puzzles in "squeeze" resolutions
  Fix uniboard layout fallback for old browsers
  improve head contours in voice rec button
  ...

+++ modules/forum/src/main/ForumDelete.scala
@@ -1,21 +1,19 @@
 package lila.forum
 
 import akka.stream.scaladsl.*
-import reactivemongo.akkastream.{ cursorProducer, AkkaStreamCursor }
 
 import lila.security.{ Granter as MasterGranter }
-import lila.user.{ Holder, User }
+import lila.user.User
 
 final class ForumDelete(
     postRepo: ForumPostRepo,
     topicRepo: ForumTopicRepo,
-    categRepo: ForumCategRepo,
     indexer: lila.hub.actors.ForumSearch,
     postApi: ForumPostApi,
     topicApi: ForumTopicApi,
     categApi: ForumCategApi,
     modLog: lila.mod.ModlogApi
-)(using ec: Executor, mat: akka.stream.Materializer):
+)(using Executor, akka.stream.Materializer):
 
   def post(categId: ForumCategId, postId: ForumPostId, mod: User): Funit =
     postRepo.unsafe.byCategAndId(categId, postId) flatMapz { post =>
@@ -50,13 +48,13 @@ final class ForumDelete(
 
   private def doDelete(view: PostView) =
     postRepo.isFirstPost(view.topic.id, view.post.id).flatMap {
-      case true =>
+      if _ then
         postRepo.idsByTopicId(view.topic.id) flatMap { postIds =>
           (postRepo removeByTopic view.topic.id zip topicRepo.remove(view.topic)) >>
             (categApi denormalize view.categ) >>-
             (indexer ! RemovePosts(postIds))
         }
-      case false =>
+      else
         postRepo.remove(view.post) >>
           (topicApi denormalize view.topic) >>
           (categApi denormalize view.categ) >>-
